fix(e2e): inline consent bootstrap into globalSetup so .js imports resolve
Jest resolves `globalSetup` outside the worker transform pipeline, and
ts-jest's `moduleNameMapper` (which strips `.js` extensions for the
ESM source layout used everywhere else) does not apply at that layer.
The previous split between `helpers/legal-consent.ts` and
`global-setup.ts` therefore failed at runtime with
`Cannot find module './helpers/legal-consent.js'` — the on-disk file
is `.ts`, but the compiled require call still asks for `.js`.

Inline the bootstrap into `tests/e2e/global-setup.ts` so the file has
zero relative imports, and re-export `bootstrapLegalConsent` for the
unit suite (which IS subject to `moduleNameMapper` and resolves the
`.js` import normally).

Verified locally:
- `pnpm jest tests/unit/legal-consent.test.ts` — 6/6 pass
- `pnpm test:unit` — 124/125 pass (1 pre-existing skip)
- `TEST_ENVIRONMENT=invalid_env npx jest --config tests/jest.e2e.config.cjs ... -t "nonexistent"` — globalSetup logs the expected warning and exits cleanly (was previously throwing a module-resolution error)
